mcp-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
PyJWT==2.8.0
cryptography==41.0.7
aiofiles==23.2.1
httpx==0.25.1
sentry-sdk[fastapi]==1.38.0